% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_outliers.R
\name{clean_outliers}
\alias{clean_outliers}
\title{Clean up outlier genes in null population}
\usage{
clean_outliers(dds, adaZmat, zcutoff, rand_seed = 19951126)
}
\arguments{
\item{dds}{a DESeqDataSet object containing the input data}

\item{adaZmat}{z-score matrix corresponding to the data in \code{dds}, produced by \code{\link{fit_main_population}} function.}

\item{zcutoff}{z-score cutoff to define the outliers as compared with the inlier population (main/null population).}

\item{rand_seed}{random seed}
}
\value{
an updated DESeqDataSet objective with all outliers in the null population replaced
\describe{
\item{\code{dds}}{updated \code{dds}}
}
}
\description{
This function uses the z-scores produced by robust main population fitting to identify the outlier conditions of each gene, and replace them with imputed inlier values.
This is a function for internal use only.
}
\author{
Xuhang Li
}
