% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_dependent_DE.R
\name{control_dependent_DE}
\alias{control_dependent_DE}
\title{Perform control dependent Differentially Expression (DE) analysis}
\usage{
control_dependent_DE(dds)
}
\arguments{
\item{dds}{A dds object from DEseq2}
}
\value{
A list of DE results (each element corresponds to the DE result of one condition, indexed by the condID):
\describe{
\item{\code{DE_table}}{The data frame containing DE results from DEseq2 for one condition.}
}
}
\description{
This is a wrapper function to perform DE analysis in the cannonical way (control-dependent) using DEseq2.
This function is designed for internal use.
}
\author{
Xuhang Li
}
