% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countTable}
\alias{countTable}
\title{An WPS dataset: count table}
\format{
A gene-by-sample dataframe of read counts
}
\usage{
data("countTable")
}
\value{
The corresponding dataframe object
}
\description{
A gene-by-sample count table that includes 10 genes across 301 conditions based on a real WPS data
}
\keyword{datasets}
