% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dds}
\alias{example_dds}
\title{An example formated DESeqDataSet object for testing control-independent DE analysis}
\format{
An example DESeqDataSet object
}
\usage{
data("example_dds")
}
\value{
The corresponding DESeqDataSet object
}
\description{
To run control-independent DE analysis individually, one should format an input DESeqDataSet with covariates included in the example:
\describe{
\item{\code{RNAi}}{covariate describing treatment (condition) of samples. A 'control' condition is not required but this covariate should be the last term in the design formula and is being tested for by default.}
\item{\code{batchLabel}}{(optional) covariate describing experimental batches and can be skipped.}
\item{\code{plate2}}{(optional) covariate indicating if the sample is cultured in the second 96-well plate. Only for WPS experiments that involves plate 2 confounding effects.}
}
}
\keyword{datasets}
