% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_stat_table}
\alias{example_stat_table}
\title{Example Statistic Table}
\format{
A data frame with 10 rows (genes) and 300 columns (conditions).
}
\usage{
data("example_stat_table")
}
\value{
The corresponding DataFrame object
}
\description{
A DataFrame object containing an example test statistic matrix. This dataset includes 10 genes across 300 conditions.
}
\keyword{datasets}
