% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_control_outliers.R
\name{find_control_outliers}
\alias{find_control_outliers}
\title{Find control-outlier genes for each library}
\usage{
find_control_outliers(
  ctr_dep_DE_res,
  libs,
  pcutoffs,
  freqCutoff,
  FCtype = "log2FoldChange_raw"
)
}
\arguments{
\item{ctr_dep_DE_res}{A list of control-dependent DE results}

\item{libs}{A character array of library IDs in the dataset}

\item{pcutoffs}{A single value or a numerical array of p-outlier cutoffs}

\item{freqCutoff}{A frequency cutoff to define the core set of control outlier gene. Default is 25\% of total number of libraries.}

\item{FCtype}{A string defining the column name of logFC in the DE result table. Only for internal use.}
}
\value{
A list of the control-outliers:
\describe{
\item{\code{genelist_all_cutoffs}}{A list of control outlier genes in each library. When \code{pcutoffs} has more than one values, this list is first grouped by different cutoffs and then by each library.}
\item{\code{coresetTbl}}{A data frame showing the name and frequency (i.e., appearing in # of libraries) of core control-outlier genes.}
}
}
\description{
This function identifies the control-outlier genes in each WPS library based on the results of control-dependent DE analysis.
This function is only designed for internal use.
}
\author{
Xuhang Li
}
