% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitEmpiricalNull.R
\name{fit_empirical_null}
\alias{fit_empirical_null}
\title{Fit the empirical null of DE test statistics}
\usage{
fit_empirical_null(statTbl, display)
}
\arguments{
\item{statTbl}{A gene-by-condition matrix of the DE test statistic (Wald statistic) from DEseq2}

\item{display}{Whether or not display the fitting of genes with heavy tails (default is FALSE)}
}
\value{
A list of the fitting results:
\describe{
\item{\code{p_mat}}{The matrix of empirical p-values (gene-by-conditions)}
\item{\code{nulls}}{A data frame for the fitted mean and standard deviations for each gene}
\item{\code{not_fit}}{Genes that were not fitted because the number of non-NA test statistics is fewer than 100 (theoratical null is used)}
\item{\code{qualityMetrics}}{A list of fitting quality metrics, including the genes whose test statistic distribution is heavy tailed (these tails were trimmed prior to fitting) and number of trimmed conditions for the fitting of each gene}
}
}
\description{
Fit the gene-by-gene empricial null distribution using an input test statistic matrix
}
\examples{
data(abc)
result <- fit_emoirical_null(XYZ)
}
\author{
Xuhang Li
}
