% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_main_population.R
\name{fit_main_population}
\alias{fit_main_population}
\title{Perform the robust fitting of the main population of each gene}
\usage{
fit_main_population(dds)
}
\arguments{
\item{statTbl}{A gene-by-condition matrix of the DE test statistic (Wald statistic) from DEseq2}

\item{display}{Whether or not display the fitting of genes with heavy tails (default is FALSE)}
}
\value{
A list of the fitting results:
\describe{
\item{\code{p_mat}}{The matrix of empirical p-values (gene-by-conditions)}
\item{\code{nulls}}{A data frame for the fitted mean and standard deviations for each gene}
\item{\code{not_fit}}{Genes that were not fitted because the number of non-NA test statistics is fewer than 100 (theoratical null is used)}
\item{\code{qualityMetrics}}{A list of fitting quality metrics, including the genes whose test statistic distribution is heavy tailed (these tails were trimmed prior to fitting) and number of trimmed conditions for the fitting of each gene}
}
}
\description{
To be updated!
}
\examples{
data(WPS_example_data)
result <- WPS_DE(countTable, metaDataTable)
}
\references{
This algorithm uses AdaTiSS published in Wang, Meng, Lihua Jiang, and Michael P. Snyder. "AdaTiSS: a novel data-Ada ptive robust method for identifying Ti ssue S pecificity S cores." Bioinformatics 37.23 (2021): 4469-4476.
}
\author{
Xuhang Li
}
