% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metaDataTable}
\alias{metaDataTable}
\title{An WPS dataset: metadata table}
\format{
A dataframe object for metadata information.
}
\usage{
data("metaDataTable")
}
\value{
The corresponding dataframe object
}
\description{
A meta-data table for the 301 samples in the example count table. WPS meta-data table should contain the following columns:
\describe{
\item{\code{sampleID}}{column that corresponds to the column names in \code{countTable}.}
\item{\code{covTreatment}}{column that is the covariate for the treatment (condition) and must have a 'control' condition.}
\item{\code{covBatch}}{column that is the covariate for batches, which by default is the replicate batch.}
\item{\code{libID}}{column that is ID of sequencing libraries. WPS DE analysis is first conducted at individual sequencing library level thus this ID is used to match samples pooled in the same library.}
\item{\code{plate2}}{column that is a logical value indicating if the sample is cultured in the second 96-well plate. This column is only needed for WPS experiments that involves plate 2 confounding effects.}
}
}
\keyword{datasets}
