% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_DE_result.R
\name{combine_DE_result}
\alias{combine_DE_result}
\title{Combine the control-dependent and -independent DE results}
\usage{
combine_DE_result(
  ctr_dep_DE_res,
  ctr_indep_DE_res,
  libs,
  cutoffs,
  gene2rmList,
  FCtype = "log2FoldChange_raw"
)
}
\arguments{
\item{ctr_dep_DE_res}{A list of data frames that are control-dependent DE results for each condition.}

\item{ctr_indep_DE_res}{A list of data frames that are control-independent DE results for each condition.}

\item{libs}{A character array of library IDs.}

\item{cutoffs}{A single value or a numerical array of p-outlier cutoffs}

\item{gene2rmList}{A list of control outlier genes for each library, identified by \code{find_control_outliers} function.}

\item{FCtype}{A string defining the column name of logFC in the DE result table. Only for internal use.}
}
\value{
A list of clean DE results with each p-outlier cutoff.
}
\description{
This function combines the DE results from both control-dependent and -independent DE analysis to resolve false positives related to control outlier genes.
This function is for internal-use only.
}
\author{
Xuhang Li
}
