% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_dependent_DE.R
\name{control_dependent_DE}
\alias{control_dependent_DE}
\title{Perform control dependent Differentially Expression (DE) analysis}
\usage{
control_dependent_DE(dds)
}
\arguments{
\item{dds}{A DESeqDataSet object defined in \code{DESeq2} package}
}
\value{
A list of DE results (each element corresponds to the DE result of one condition, indexed by the condID).
}
\description{
This is a wrapper function to perform DE analysis in the cannonical way (control-dependent) using DEseq2.
This function is designed for internal use.
}
\author{
Xuhang Li
}
