% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_independent_DE.R
\name{control_independent_DE}
\alias{control_independent_DE}
\title{Control-independent Differential Expression (DE) analysis}
\usage{
control_independent_DE(dds_ori, adaZmat, zcutoff = 2.5)
}
\arguments{
\item{dds_ori}{A DESeqDataSet objective that contains the input data. When this function is used alone, the \code{dds} object should be constructed following the metadata format of WPS dataset (see examples).}

\item{adaZmat}{A gene-by-condition matrix of z-scores obtained through robust fitting of the main population. See \code{fit_main_population} function for details.}

\item{zcutoff}{z-score cutoff to define the outlier conditions with respect to the main population.}
}
\value{
A list of DE results for each condition.
}
\description{
Perform control-independent DE analysis based on a null population defined by robust fitting of the main population of each individual gene expression.
}
\examples{
data(ctr_indep_DE_example_data)
result <- control_independent_DE(dds_ori, adaZmat)

# To generate the z-score matrix, run:
adaZmat <- fit_main_population(dds_ori)
}
\author{
Xuhang Li
}
