% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_control_outliers.R
\name{find_control_outliers}
\alias{find_control_outliers}
\title{Find control-outlier genes for each library}
\usage{
find_control_outliers(
  ctr_dep_DE_res,
  libs,
  pcutoffs,
  freqCutoff,
  FCtype = "log2FoldChange_raw"
)
}
\arguments{
\item{statTbl}{A gene-by-condition matrix of the DE test statistic (Wald statistic) from DEseq2}

\item{display}{Whether or not display the fitting of genes with heavy tails (default is FALSE)}
}
\value{
A list of the fitting results:
\describe{
\item{\code{p_mat}}{The matrix of empirical p-values (gene-by-conditions)}
\item{\code{nulls}}{A data frame for the fitted mean and standard deviations for each gene}
\item{\code{not_fit}}{Genes that were not fitted because the number of non-NA test statistics is fewer than 100 (theoratical null is used)}
\item{\code{qualityMetrics}}{A list of fitting quality metrics, including the genes whose test statistic distribution is heavy tailed (these tails were trimmed prior to fitting) and number of trimmed conditions for the fitting of each gene}
}
}
\description{
Perform the two-pronged WPS DE analysis with an input dataset consistent with WPS data structure.
}
\examples{
data(WPS_example_data)
result <- WPS_DE(countTable, metaDataTable)
}
\author{
Xuhang Li
}
