% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_main_population.R
\name{fit_main_population}
\alias{fit_main_population}
\title{Perform the robust fitting of the main population of each gene}
\usage{
fit_main_population(dds)
}
\arguments{
\item{dds}{A DESeqDataSet objective that contains the input data. When this function is used alone, the \code{dds} object should be constructed following the metadata format of WPS dataset (see examples).}
}
\value{
A matrix of fitted z-scores for each gene in each sample.
}
\description{
This is a wrapper function that uses \code{AdaTiSS} algorithm (Wang et al. 2021) to fit the main population of the expression levels of a gene.
The mean and variance of this main population are used to calculate z-scores for each gene in each condition, which are further used in the control-independent
DE analysis to remove outliers of the main population.
}
\examples{
data(example_dds)
adaZmat <- fit_main_population(example_dds)
}
\references{
This function uses AdaTiSS algorithm published by Wang et al.

\cite{Wang, Meng, Lihua Jiang, and Michael P. Snyder. "AdaTiSS: a novel data-Ada ptive robust method for identifying Ti ssue S pecificity S cores." Bioinformatics 37.23 (2021): 4469-4476.}
}
\author{
Xuhang Li
}
